Move styles to own file
